server/pyproject.toml
[project]
name = "smello-server"
version = "0.1.0"
description = "A local web dashboard for inspecting outgoing HTTP requests from your code"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.0",
    "tortoise-orm>=0.22.0",
    "aiosqlite>=0.20.0",
]

[project.scripts]
smello-server = "smello_server.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/smello_server"]
